requirements.txt
psycopg2-binary
chromadb
openai
crewai
reportlab
